requirements.txt
Flask>=3.0,<4
gunicorn>=22,<23
SQLAlchemy>=2.0,<3
psycopg2-binary>=2.9,<3
requests>=2.31,<3
requests-oauthlib>=2.0,<3
feedparser>=6.0,<7
python-dotenv